[hotfix] Move TaskManagersHandler to rest.handler.taskmanager and messages to rest.messages.taskmanager
Move TaskManager messages to rest.messages.taskmanager

Move TaskManager message tests to rest.messages.taskmanager
